[Fix] print format

+++ 0x1E-search_algorithms/1-binary.c
@@ -44,7 +44,7 @@ int binary_search(int *array, size_t size, int value)
 		else if (array[middle] > value)
 			right = middle - 1;
 		else
-			return (middle);
+			return ((int)middle);
 	}
 	return (-1);
 }
